Comprueba si el entorno existe, si no... lo crea

+++ config/config.sh
@@ -1,16 +1,19 @@
 #!/bin/bash
 
+cd ../Back/Python_WS/
 
-cd ..
+entorno="/entorno"
 
-cd Back/Python_WS/
+if [ -e $entorno ]; then 
+    printf "\nEl entorno ya existe y esta configurado"
+else	
+    printf "\nConfigurando entorno de python..."		
+    python -m venv entorno 
 
+    source entorno/bin/activate
 
-python -m venv entorno 
-
-source entorno/bin/activate
-
-pip install peewee
-pip install uvicorn 
-pip install fastapi
-pip install PyMySQL
+    pip install peewee
+    pip install uvicorn 
+    pip install fastapi
+    pip install PyMySQL
+fi
